Refactor DDLogMessage, DDLoggerNode and internal methods

+++ Classes/Extensions/DDContextFilterLogFormatter.h
@@ -39,12 +39,12 @@
 
 - (instancetype)init NS_DESIGNATED_INITIALIZER;
 
-- (void)addToWhitelist:(int)loggingContext;
-- (void)removeFromWhitelist:(int)loggingContext;
+- (void)addToWhitelist:(NSUInteger)loggingContext;
+- (void)removeFromWhitelist:(NSUInteger)loggingContext;
 
 @property (readonly, copy) NSArray *whitelist;
 
-- (BOOL)isOnWhitelist:(int)loggingContext;
+- (BOOL)isOnWhitelist:(NSUInteger)loggingContext;
 
 @end
 
@@ -59,11 +59,11 @@
 
 - (instancetype)init NS_DESIGNATED_INITIALIZER;
 
-- (void)addToBlacklist:(int)loggingContext;
-- (void)removeFromBlacklist:(int)loggingContext;
+- (void)addToBlacklist:(NSUInteger)loggingContext;
+- (void)removeFromBlacklist:(NSUInteger)loggingContext;
 
 @property (readonly, copy) NSArray *blacklist;
 
-- (BOOL)isOnBlacklist:(int)loggingContext;
+- (BOOL)isOnBlacklist:(NSUInteger)loggingContext;
 
 @end
